requirements.txt
# Core ML
xgboost>=2.0.0
scikit-learn>=1.4.0
statsmodels>=0.14.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0

# Data fetching
requests>=2.31.0
pybaseball>=2.3.0

# Evaluation & visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
tqdm>=4.66.0

# Notifications (Phase 10 — optional)
anthropic>=0.40.0
twilio>=9.0.0
